Implement `ldexp` and use it in `ratio_to_f64`
This eliminates errors where we attempt a multiplication
by 2^-x, but that 2^-x underflows.

The `ldexp` implementation strategy is mostly to extract
the exponent, modify it, then reinsert it. To avoid
dealing with subnormal numbers at the bit level, we
shift any subnormal numbers into normal range before
that operation. This causes no loss of precision.